feat(cli): Write CLI
This introduces the entirety of notification querying from the CLI
including a custom lexer & parser for filtering notifications. The
query language currently supports the following OPS and Functions
- & / and / AND
- | / or / OR
- ! / not / NOT
- { attribue=value } * Dbus attribute search
- (...) * Grouping
- X..Y * Every notification betweenm X & Y inclusive
- unread() * Every unread notification (not marked closed)
- last(X) * Gets the last X number of notifications
- page(offset, per_page) * A pageinator

it also makes some minor modifications to the DB & rust struct to have
better parity, particularly in the case of hints as `key` is a sql
keyword, so it's been renamed to `k`

The notification server can now be started by running
`sprinkles notification server` and it will create & store a database at
`~/.local/share/sprinkles/sprinkles.db` instead of `test.db`

As this is a officially working product I will be marking this as our
first working alpha version `0.0.1`. The Quickshell GUI for this should
hopefully be coming soon